requirements.txt
requests~=2.32.3
ultralytics~=8.3.143
torch~=2.7.0
torchvision~=0.22.0
pillow~=11.1.0
transformers~=4.52.3
MyApplication~=0.1.0
opencv-python~=4.11.0.86
pytesseract~=0.3.13
fastapi~=0.115.11
uvicorn~=0.32.0
numpy~=2.2.6
langchain-google-genai~=2.1.4
python-dotenv~=1.1.0
easyocr~=1.7.2
langchain-core~=0.3.61
mistralai~=1.7.1
pydantic~=2.11.5
langchain~=0.3.25